Refactor ExerciseViewController Inject ExerciseViewModel directly into ExerciseItemCell Move ExerciseItemViewModel into a separate file
